requirements.txt
tango-python
python-dotenv
pandas
requests
